Update program 05.py

+++ ProgramBatch1/program 05.py
@@ -1,4 +1,4 @@
-num1 = int(input("Enter first number: "))
-num2 = int(input("Enter second number: "))
+num1 = float(input("Enter first number: "))
+num2 = float(input("Enter second number: "))
 print("quotient:", num1 / num2)
